+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -10,6 +10,9 @@ Released 2021-Nov-19
   `FormatException` if it fails to parse any of the supported environment
   variables.
 
+* Added `HttpClientFactory` option
+  ([#2654](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2654))
+
 ## 1.2.0-beta1
 
 Released 2021-Oct-08